流れでImGuiを実装
GMの資料が大変参考になった。
次あたりでデバッグ関数の登録などなどをしてみたい。
というのは我慢してSceneManagerを完成に近づけつつ試せたらと思う。

+++ LightThrough/SourceFiles/DX3D/Include/DX3D/Math/Point.h
@@ -21,7 +21,7 @@ namespace dx3d {
 	class Point {
 	public:
 		Point() : x(0), y(0) {}
-		Point(int _x, int _y) : x(_x), y(_y) {}
+		Point(int _x, int _y) : x(static_cast<f32>(_x)), y(static_cast<f32>(_y)) {}
 		Point(f32 _x, f32 _y) : x(_x), y(_y) {}
 		Point(const Point& _point) : x(_point.x), y(_point.y) {}
 		~Point() {}
